Fixes in package.json file for BIN and Require support.
In order to support using the package with commands such as require you need to define the entry point
using the main key in package.json, however to use it as a command in the PATH you must add a shebang to it
and define it using the bin key.
A problem arises because the shebang makes the entry point technically invalid Javascript and as such it should only be
added to the specific script that is included in the PATH and not on the main Node entry point that is meant to be used with Require.